+++ overlays/nix/flake.nix
@@ -1,30 +1,22 @@
 {
+  inputs.nix.url = github:NixOS/nix;
   inputs.nix.inputs.nixpkgs.follows = "nixpkgs";
-  inputs.inheritedNix.inputs.nixpkgs.follows = "nixpkgs";
 
   inputs.flake-utils.url = github:numtide/flake-utils;
 
-  outputs = { self, nixpkgs, inheritedNix, nix, flake-utils }: {
+  outputs = { self, nixpkgs, nix, flake-utils }: {
     overlay = final: prev: {
-      nix = prev.nix.overrideAttrs (old:
-        # `inheritedNix` and `nix` should always have the same NAR hash
-        # except when `inheritedNix` is overridden using `--override-input nix ...`
-        if (inheritedNix.narHash != nix.narHash) then {
-          version = "${old.version}-dirtiest";
+      nix = nix.defaultPackage.${prev.system}.overrideAttrs (old: {
+        version = "${old.version}-dirtier";
 
-          src = inheritedNix;
-        } else {
-          version = "${old.version}-dirtier";
-
-          patches = old.patches ++ [
-            (prev.fetchpatch {
-              name = "add-dirtyRev-and-dirtyShortRev-to-fetchGit.patch";
-              url = "https://patch-diff.githubusercontent.com/raw/NixOS/nix/pull/5385.patch";
-              sha256 = "sha256-50qV1srrwbCICgY9XRvX7EHpU1ZtdXE8jkCgy5QeMh0=";
-            })
-          ];
-        }
-      );
+        patches = (old.patches or [ ]) ++ [
+          (prev.fetchpatch {
+            name = "add-dirtyRev-and-dirtyShortRev-to-fetchGit.patch";
+            url = "https://patch-diff.githubusercontent.com/raw/NixOS/nix/pull/5385.patch";
+            sha256 = "sha256-KVXTpSRIsgvyisJeWMan1fUWgyGpx17mUGZjE+QVPbI=";
+          })
+        ];
+      });
     };
   } // (
     flake-utils.lib.eachDefaultSystem (system: {